Prepare assembly admin ability for other roles
